Basic Render Debug Information
Add a key that dumps out the last frames rendering details
* Time/Vertices in each Renderer
* Time/Vertices in each Shader
* For each render pass (reflection, refraction and main)

Enabled asserts on Release as that's where most of my development
happens
Fix some compiler warnings

Add EntityManager to help manage Entites in the game
Start breaking main into smaller pieces

+++ Terrain/HeightsGenerator.h
@@ -34,10 +34,10 @@ public:
 
 	float GenerateHeight(int aX, int aZ)
 	{
-		int x = myXOffset + aX;
-		int z = myZOffset + aZ;
+		float x = static_cast<float> (myXOffset + aX);
+		float z = static_cast<float>(myZOffset + aZ);
 
-		float total = GetInterpolatedNoise(x / 32.0f, z / 32.0f) * AMPLITUTE * 10.0;
+		float total = GetInterpolatedNoise(x / 32.0f, z / 32.0f) * AMPLITUTE * 10.0f;
 		total += GetInterpolatedNoise(x / 8.0f, z / 8.0f) * AMPLITUTE * 2.0f;
 		total += GetInterpolatedNoise(x / 2.0f, z / 2.0f) * AMPLITUTE / 6.0f;
 		total += GetInterpolatedNoise(x, z) * AMPLITUTE / 12.0f;
